+++ cs/remote/test/FASTER.remote.test/FixedLenBinaryTests.cs
@@ -37,34 +37,5 @@ namespace FASTER.remote.test
             session.Read(20, userContext: 23 * 3);
             session.CompletePending(true);
         }
-
-        [Test]
-        public void SubscribeTest()
-        {
-            var session = client.GetSession();
-            var subSession = client.GetSession();
-
-            subSession.SubscribeKV(10);
-            subSession.CompletePending(true);
-            session.Upsert(10, 23);
-            session.CompletePending(true);
-
-            System.Threading.Thread.Sleep(1000);
-        }
-
-        [Test]
-        public void PrefixSubscribeTest()
-        {
-            var session = client.GetSession();
-            var subSession = client.GetSession();
-
-            subSession.PSubscribeKV(10);
-            subSession.CompletePending(true);
-            session.Upsert(10, 23);
-            session.CompletePending(true);
-
-            System.Threading.Thread.Sleep(1000);
-        }
-
     }
 }
