+++ cs/remote/test/FASTER.remote.test/VarLenBinaryTests.cs
@@ -39,50 +39,5 @@ namespace FASTER.remote.test
             session.Read(key, userContext: key.Span[0]);
             session.CompletePending(true);
         }
-
-        [Test]
-        public void SubscribeTest()
-        {
-            Random r = new Random(23);
-            var session = client.GetSession();
-            var subSession = client.GetSession();
-            var key = new Memory<int>(new int[2 + r.Next(50)]);
-            var value = new Memory<int>(new int[1 + r.Next(50)]);
-            key.Span[0] = r.Next(100);
-            key.Span[1] = value.Length;
-            value.Span.Fill(key.Span[0]);
-
-            subSession.SubscribeKV(key);
-            subSession.CompletePending(true);
-            session.Upsert(key, value);
-            session.CompletePending(true);
-
-            System.Threading.Thread.Sleep(1000);
-        }
-
-        [Test]
-        public void PSubscribeTest()
-        {
-            Random r = new Random(23);
-            var session = client.GetSession();
-            var subSession = client.GetSession();
-            var key = new Memory<int>(new int[2 + r.Next(50)]);
-            var value = new Memory<int>(new int[1 + r.Next(50)]);
-            int randomNum = r.Next(100);            
-            key.Span[0] = randomNum;
-            key.Span[1] = value.Length;
-            value.Span.Fill(key.Span[0]);
-
-            var upsertKey = new Memory<int>(new int[100]);
-            upsertKey.Span[0] = randomNum;
-            upsertKey.Span[1] = value.Length;
-
-            subSession.PSubscribeKV(key);
-            subSession.CompletePending(true);
-            session.Upsert(upsertKey, value);
-            session.CompletePending(true);
-
-            System.Threading.Thread.Sleep(1000);
-        }
     }
 }
