+++ cs/remote/test/FASTER.remote.test/VarLenServer.cs
@@ -13,14 +13,17 @@ namespace FASTER.remote.test
         readonly FasterKV<SpanByte, SpanByte> store;
         readonly SpanByteFasterKVProvider provider;
 
-        public VarLenServer(string folderName, string address = "127.0.0.1", int port = 33278)
+        public VarLenServer(string folderName, string address = "127.0.0.1", int port = 33278, bool useBroker = false)
         {
             this.folderName = folderName;
             GetSettings(folderName, out var logSettings, out var checkpointSettings, out var indexSize);
 
             store = new FasterKV<SpanByte, SpanByte>(indexSize, logSettings, checkpointSettings);
+            
+            SubscribeKVBroker<SpanByte, SpanByte, IKeySerializer<SpanByte>> broker = null;
 
-            var broker = new SubscribeKVBroker<SpanByte, SpanByte, IKeySerializer<SpanByte>>(new SpanByteKeySerializer());
+            if (useBroker)
+                broker = new SubscribeKVBroker<SpanByte, SpanByte, IKeySerializer<SpanByte>>(new SpanByteKeySerializer());
 
             // Create session provider for VarLen
             provider = new SpanByteFasterKVProvider(store, broker);
